backend/requirements.txt
django==4.2.25
psycopg2-binary==2.9.10
python-dotenv==1.0.0
djangorestframework